Split permission types
