+++ packages/types/src/discordeno.ts
@@ -1,4 +1,3 @@
-import type { DiscordRole } from './discord.js'
 import type {
   ApplicationFlags,
   DiscordApplicationEventWebhookStatus,
@@ -57,9 +56,10 @@ import type {
   DiscordReactionType,
   MessageFlags,
 } from './discord/messages.js'
+import type { DiscordRole, PermissionStrings } from './discord/permissions.js'
 import type { DiscordPollAnswer, DiscordPollLayoutType, DiscordPollMedia } from './discord/polls.js'
 import type { DiscordWebhookEventType } from './discord/webhookEvents.js'
-import type { BigString, Camelize, Localization, PermissionStrings } from './shared.js'
+import type { BigString, Camelize, Localization } from './shared.js'
 
 /** https://discord.com/developers/docs/resources/channel#create-message-jsonform-params */
 export interface CreateMessageOptions {
